feat: cleaned lib exposure from core, exposing only usefull traits
