+++ cli/src/output/formats/json.rs
@@ -3,11 +3,11 @@ use std::io::Write;
 
 use crate::output::displayer::error::IntoDisplayerError;
 use crate::output::displayer::{Displayer, DisplayerError};
-use joule_profiler_core::profiler::types::Iteration;
-use joule_profiler_core::sensor::Sensor;
-use joule_profiler_core::util::file::{
+use joule_profiler_core::fs::{
     create_file_with_user_permissions, default_iterations_filename, get_absolute_path,
 };
+use joule_profiler_core::sensor::Sensor;
+use joule_profiler_core::types::Iteration;
 use serde_json::json;
 
 type Result<T> = std::result::Result<T, DisplayerError>;
