+++ core/Cargo.toml
@@ -14,6 +14,7 @@ serde.workspace = true
 tokio.workspace = true
 thiserror.workspace = true
 regex = "1.12.2"
+derive_builder = "0.20.2"
 
 [dev-dependencies]
 tempfile = "3"
