+++ core/src/util/fs.rs
@@ -7,6 +7,23 @@ use std::{
 
 use crate::util::time::get_timestamp_millis;
 
+/// Get the absolute path of a file.
+pub fn get_absolute_path(filename: &str) -> Result<String, std::io::Error> {
+    let path = PathBuf::from(filename);
+    let absolute_path = if path.is_absolute() {
+        path
+    } else {
+        env::current_dir()?.join(path)
+    };
+
+    Ok(absolute_path.display().to_string())
+}
+
+/// Generates a default filename for iteration data.
+pub fn default_iterations_filename(ext: &str) -> String {
+    format!("data{}.{}", get_timestamp_millis(), ext)
+}
+
 const ROOT_UID_ENV_VAR: &str = "SUDO_UID";
 const ROOT_GID_ENV_VAR: &str = "SUDO_GID";
 const URW_GRW_OR_PERMS: u32 = 0o664;
@@ -32,21 +49,4 @@ pub fn create_file_with_user_permissions(path: &str) -> std::io::Result<File> {
     chown(path, uid, gid)?;
 
     Ok(file)
-}
-
-/// Get the absolute path of a file.
-pub fn get_absolute_path(filename: &str) -> Result<String, std::io::Error> {
-    let path = PathBuf::from(filename);
-    let absolute_path = if path.is_absolute() {
-        path
-    } else {
-        env::current_dir()?.join(path)
-    };
-
-    Ok(absolute_path.display().to_string())
-}
-
-/// Generates a default filename for iteration data.
-pub fn default_iterations_filename(ext: &str) -> String {
-    format!("data{}.{}", get_timestamp_millis(), ext)
 }
