+++ core/src/util/mod.rs
@@ -1,2 +1,2 @@
-pub mod file;
+pub mod fs;
 pub mod time;
